Do not mention conda skeleton in README.
While nice in theory, conda skeleton produces a recipe that will need to be modified a lot (it does not use template variables, puts all dependencies in both build and run, uses md5 hash, uses build scripts for simple python packages, ...). It might be a better idea to just start with the example recipe and change where change is needed.

Most of the comments in #2289 referred to the lines generated by `conda skeleton`.